refactor: move nonprofit functionality to dedicated app
Created new nonprofit_orgs dedicated Django app
Moved NonprofitOrg models from note to nonprofit_orgs
Moved EndaomentService and related functionality
Initial migration depends on 0028_payment because of Foreign Key
Updated URL routes to use /api/nonprofit/ endpoints
Cleaned up note app, removing all nonprofit references
Moved test coverage to the new app
