+++ src/nonprofit_orgs/urls.py
@@ -1,8 +1,10 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
+from nonprofit_orgs.views import NonprofitOrgViewSet
+
 router = DefaultRouter()
-# No endpoints defined
+router.register(r"orgs", NonprofitOrgViewSet, basename="nonprofit-orgs")
 
 urlpatterns = [
     path("", include(router.urls)),
