+++ src/note/models.py
@@ -1,4 +1,3 @@
-from note.related_models.nonprofit_org_model import NonprofitFundraiseLink, NonprofitOrg
 from note.related_models.note_model import Note, NoteContent
 from note.related_models.note_template_model import NoteTemplate
 
@@ -7,6 +6,4 @@ __all__ = [
     "Note",
     "NoteContent",
     "NoteTemplate",
-    "NonprofitOrg",
-    "NonprofitFundraiseLink",
 ]
