+++ src/researchhub/urls.py
@@ -207,6 +207,7 @@ urlpatterns = [
     path("api/permissions/", researchhub.views.permissions, name="permissions"),
     path("api/search/", include(search.urls)),
     path("api/note/", include("note.urls")),
+    path("api/nonprofit/", include("nonprofit_orgs.urls")),
     path(
         "auth/google/yolo/callback/",
         oauth.views.google_yolo_callback,
